requirements.txt
torch
torchvision
numpy
opencv-python
Pillow
tqdm
tensorboard
ultralytics
deep-sort-realtime